// src/enums/JudgeResultEnum.ts
export enum JudgeResultEnum {
  ACCEPTED = "ACCEPTED",
  WRONG_ANSWER = "WRONG_ANSWER",
  TIME_LIMIT_EXCEEDED = "TIME_LIMIT_EXCEEDED",
  MEMORY_LIMIT_EXCEEDED = "MEMORY_LIMIT_EXCEEDED",
  RUNTIME_ERROR = "RUNTIME_ERROR",
  COMPILATION_ERROR = "COMPILATION_ERROR",
  PRESENTATION_ERROR = "PRESENTATION_ERROR",
  CLASS_NAME_MISMATCH = "CLASS_NAME_MISMATCH",
}